+++ tests/Unit/Delivery/ImageOptionsTest.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @copyright 2015 Contentful GmbH
+ * @copyright 2015-2016 Contentful GmbH
  * @license   MIT
  */
 
@@ -314,6 +314,225 @@ class ImageOptionsTest extends \PHPUnit_Framework_TestCase
 
         $this->assertSame('jpg', $options->getFormat());
         $this->assertSame('fm=jpg&fl=progressive', $options->getQueryString());
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setResizeFit
+     * @covers Contentful\Delivery\ImageOptions::getResizeFit
+     */
+    public function testGetSetResizeFit()
+    {
+        $options = new ImageOptions;
+        $options->setResizeFit('pad');
+
+        $this->assertEquals('pad', $options->getResizeFit());
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setResizeFit
+     *
+     * @expectedException \InvalidArgumentException
+     */
+    public function testSetResizeFitInvalid()
+    {
+        $options = new ImageOptions;
+        $options->setResizeFit('invalid');
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setResizeFit
+     * @covers Contentful\Delivery\ImageOptions::getQueryString
+     */
+    public function testQueryResizeFit()
+    {
+        $options = new ImageOptions;
+        $options->setResizeFit('pad');
+
+        $this->assertSame('fit=pad', $options->getQueryString());
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setResizeFocus
+     * @covers Contentful\Delivery\ImageOptions::getResizeFocus
+     */
+    public function testGetSetResizeFocus()
+    {
+        $options = new ImageOptions;
+        $options->setResizeFocus('top');
+
+        $this->assertEquals('top', $options->getResizeFocus());
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setResizeFocus
+     *
+     * @expectedException \InvalidArgumentException
+     */
+    public function testSetResizeFocusInvalid()
+    {
+        $options = new ImageOptions;
+        $options->setResizeFocus('invalid');
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setResizeFit
+     * @covers Contentful\Delivery\ImageOptions::setResizeFocus
+     * @covers Contentful\Delivery\ImageOptions::getQueryString
+     */
+    public function testQueryResizeFocus()
+    {
+        $options = new ImageOptions;
+        $options->setResizeFit('thumb');
+        $options->setResizeFocus('top');
+
+        $this->assertSame('fit=thumb&f=top', $options->getQueryString());
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setResizeFocus
+     * @covers Contentful\Delivery\ImageOptions::getQueryString
+     */
+    public function testQueryResizeFocusIgnoredWithoutFit()
+    {
+        $options = new ImageOptions;
+        $options->setResizeFocus('top');
+
+        $this->assertSame('', $options->getQueryString());
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setRadius
+     * @covers Contentful\Delivery\ImageOptions::getRadius
+     */
+    public function testGetSetRadius()
+    {
+        $options = new ImageOptions;
+        $options->setRadius(50.3);
+
+        $this->assertEquals(50.3, $options->getRadius());
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setRadius
+     *
+     * @expectedException \InvalidArgumentException
+     */
+    public function testSetRadiusNegative()
+    {
+        $options = new ImageOptions;
+        $options->setRadius(-13.2);
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setRadius
+     * @covers Contentful\Delivery\ImageOptions::getQueryString
+     */
+    public function testQueryRadius()
+    {
+        $options = new ImageOptions;
+        $options->setRadius(50.3);
+
+        $this->assertSame('r=50.3', $options->getQueryString());
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setBackgroundColor
+     * @covers Contentful\Delivery\ImageOptions::getBackgroundColor
+     */
+    public function testGetSetBackgroundColorSixDigits()
+    {
+        $options = new ImageOptions;
+        $options->setBackgroundColor('#a0f326');
+
+        $this->assertEquals('#a0f326', $options->getBackgroundColor());
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setBackgroundColor
+     * @covers Contentful\Delivery\ImageOptions::getBackgroundColor
+     */
+    public function testGetSetBackgroundColorThreeDigits()
+    {
+        $options = new ImageOptions;
+        $options->setBackgroundColor('#0AF');
+
+        $this->assertEquals('#0AF', $options->getBackgroundColor());
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setBackgroundColor
+     * @covers Contentful\Delivery\ImageOptions::getBackgroundColor
+     */
+    public function testGetSetBackgroundColorUpperCase()
+    {
+        $options = new ImageOptions;
+        $options->setBackgroundColor('#A0F326');
+
+        $this->assertEquals('#A0F326', $options->getBackgroundColor());
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setBackgroundColor
+     *
+     * @expectedException \InvalidArgumentException
+     */
+    public function testSetBackgroundColorTooShort()
+    {
+        $options = new ImageOptions;
+        $options->setBackgroundColor('#A0F36');
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setBackgroundColor
+     *
+     * @expectedException \InvalidArgumentException
+     */
+    public function testSetBackgroundInvalidCharacter()
+    {
+        $options = new ImageOptions;
+        $options->setBackgroundColor('#A0H326');
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setBackgroundColor
+     *
+     * @expectedException \InvalidArgumentException
+     */
+    public function testSetBackgroundNoHash()
+    {
+        $options = new ImageOptions;
+        $options->setBackgroundColor('A0F326');
+    }
+
+    /**
+     * @covers Contentful\Delivery\ImageOptions::__construct
+     * @covers Contentful\Delivery\ImageOptions::setResizeFit
+     * @covers Contentful\Delivery\ImageOptions::setBackgroundColor
+     * @covers Contentful\Delivery\ImageOptions::getQueryString
+     */
+    public function testQueryBackgroundColor()
+    {
+        $options = new ImageOptions;
+        $options->setResizeFit('pad');
+        $options->setBackgroundColor('#a0f326');
+
+        $this->assertSame('fit=pad&bg=rgb%3Aa0f326', $options->getQueryString());
     }
 
     /**