+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+* Implemented missing functionality of Contentful Images API in ImageOptions.
+
 ### Fixed
 * Logged requests were always shown as belonging to the Delivery API, even when the Preview API was used.
 * Fields not present in an Entry would lead to an error.